bot/requirements.txt
aiogram
python-dotenv
redis
psycopg2-binary
requests
